+++ CHANGELOG.md
@@ -45,6 +45,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   }
   ```
   
+
+* Smart enum property support. Just add a property and give it a list of options using the `SmartNames` property and you're good to go. They will all be strings with smart enum backing. Attributes like filtering and sorting are supported
+
+  ```yaml
+  
+    Entities:
+    - Name: Recipe
+      Features:
+      #...
+      Properties:
+      - Name: Title
+        Type: string
+        CanFilter: true
+        CanSort: true
+      - Name: Visibility
+        SmartNames:
+        - Public
+        - Friends Only
+        - Private
+        CanFilter: true
+        CanSort: true
+  ```
+